chore: move student dashboard (from modules.tsx to modules)
